requirements.txt
opencv-python
mediapipe
PyAutoGUI
SpeechRecognition
tensorflow
PyAudio
pynput
pywin32
keyboard
PyYAML